build

+++ src/server/game/AI/CreatureAI.cpp
@@ -292,14 +292,14 @@ void CreatureAI::MoveBackwardsChecks() {
 
 Creature* CreatureAI::DoSummon(uint32 entry, const Position& pos, Milliseconds despawnTime, TempSummonType summonType)
 {
-    return me->SummonCreature(entry, pos, summonType, despawnTime);
+    return me->SummonCreature(entry, pos, summonType, static_cast<uint32>(despawnTime));
 }
 
 Creature* CreatureAI::DoSummon(uint32 entry, WorldObject* obj, float radius, Milliseconds despawnTime, TempSummonType summonType)
 {
     Position pos;
     obj->GetRandomNearPosition(pos, radius);
-    return me->SummonCreature(entry, pos, summonType, despawnTime);
+    return me->SummonCreature(entry, pos, summonType, static_cast<uint32>(despawnTime));
 }
 
 Creature* CreatureAI::DoSummonFlyer(uint32 entry, WorldObject* obj, float flightZ, float radius, Milliseconds despawnTime, TempSummonType summonType)
@@ -307,5 +307,5 @@ Creature* CreatureAI::DoSummonFlyer(uint32 entry, WorldObject* obj, float flight
     Position pos;
     obj->GetRandomNearPosition(pos, radius);
     pos.m_positionZ += flightZ;
-    return me->SummonCreature(entry, pos, summonType, despawnTime);
+    return me->SummonCreature(entry, pos, summonType, static_cast<uint32>(despawnTime));
 }
